Add base Pawn piece logic
This commit implements the pawn movement logic. The overall logic
appears to be a bit jank, but future commits could possibly simplify the
logic with lambdas. I still need to learn about lambdas, however.

Also, the abstract Piece parameters now have the protected status so
that extended parameters can alter them. I feel like this is not needed,
and could be private, but I am not sure. Might be a bug with the IDE, or
a bug with my brain.

All Piece objects now need to implement canMove() and move() methods.