Add overriden move() for Pawn and fix Piece move()
The move() method for the Pawn piece needs to check the enpassant move.
The only caveat of this design is that the overall game manager would
need to check if a given pawn has double-stepped, then set that boolean
to false by calling resetDoubleStep() once the enemy moved.

Also, there was a bug in the Piece.java where the y-value was instead an
x.