Fix AIOBException on King and Knight
The position check to make sure that a given position is on the board is
now made into a static method which checks two integers. Queen.java was
altered to accommodate this change.

There shouldn't be a position object that isn't on the board. Waste of
time creating an object like that.