Add GUI base and refactor with JavaFX
A new project needed to be made with the old files transferred over. One
of the most important things that I've learned is that when starting a
gui project like this, the pom.xml needs to have the correct mainClass
tag in the openjfx plugin. Also, the module-info.java also needs to
export the package that the starting GUI file is contained.