Add image to the corresponding pieces on the GUI
For the time being, the images have been moved to the base root source
folder. For some reason, the program cannot read files that are inside
of directories.

My keyboard's S key has been acting up, so I will be taking some time
off programming to fix this. It's so damn annoying.