Add last changes to King's movement
This project was supposed to be super short and sweet, but it has
instead gone on for 2 weeks. It's time to end this. I've learned enough
about GUI, code architecture, and lambda expressions from this.
This commit's gui can result in errors. 2 commits before this one will
be the official version of this project unless I decide to work on this
more.